Refactored file structure slightly for game model.
