+++ api/game/gameHelpers/index.js
@@ -1,0 +1,7 @@
+const points = require('./pointAssignment');
+const faceoff = require('./faceoffRetrieval');
+
+module.exports = {
+  points,
+  faceoff,
+};
